+++ lib/domain/diary/confirm_diary_sheet.dart
@@ -1,4 +1,4 @@
-import 'package:pilll/components/atoms/buttons.dart';
+import 'package:pilll/components/atoms/button.dart';
 import 'package:pilll/components/page/discard_dialog.dart';
 import 'package:pilll/domain/diary_post/diary_post_page.dart';
 import 'package:pilll/entity/diary.codegen.dart';
